LTC Copier
Copyright 2017 - writed by Rojo Ambinina

This code is just a homemade webscraping algorithm, to store in a file "Litecoin_address.txt" the top 100 richest in litecoin on https://bitinfocharts.com using the module BeautifulSoup and Requests.

To use this little code, you just have to install the BeautifulSoup and Requests modules. That's all needed.